create AddAccount component
